Changed default number of passes from the TerserPlugin from 2 to 1
Changed the number of passes by the TerserPlugin from 2 to 1, which adds a few bytes to the output, but fixes the incompatibility with the new SWC minifier that is now default in Next.js.

Required updating Webpack to a version above 5.87.0, when the ability to do this was added.

Doing this created linting errors until every other package was also updated.